==> 01_read_and_merge.Rmd <==
---
title: "Gaze Following Data Reading and Merge"
author: "Gaze following Analysis Team"
date: '`r format(Sys.time(), "%a %b %d %X %Y")`'
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: yes
---


# Intro

This script reads and merges data from the multi-lab gaze-following project.  This script takes data from eye-trackers that were pre-processed by gazepath scripts written by Daan van Renswoude, and hand-coded data in the same format. Credit to the much code borrowed from mb1/mb1b scripts.

* `01_read_and_merge` reads and merges the data. The goal of this file is to create a single inclusive file that has all data from all labs. 
* `02_variable_validation` corrects errors in variables and ensures that formats are correcvt. 
* `03_exclusion` reports exclusions and creates diff files. 
* `04_confirmatory_analysis` is the set of confirmatory analyses that were preregisted

# Participant Import 


```{r setup, echo=FALSE, message=FALSE}
source("helper/common.R")
```

Data import functions are factored into a helper functions file. 

```{r}
source("helper/preprocessing_helper.R")
```
    

# Reading in all participant data.  Excluding problematic columns.  Coercing to character if any problems.

```{r}
participant_files <- dir("processed_data/participants_cleaned/", pattern = ".xlsx", all.files = FALSE)
participant_files <- paste0("processed_data/participants_cleaned/", participant_files) # Adds path to names

for (file in participant_files){
  # if the merged dataset does exist, append to it

  if (exists("pd_raw")){
    temp_dataset <- read_excel(file, col_names=TRUE) %>%
      mutate_all(as.character) %>%
      mutate_all(nc_na) %>%
      select(-(starts_with("Optional"))) %>%
      select(-(starts_with("parent")))%>%
      select(-contains("NAE")) %>%
      select(-starts_with("Bilingual")) %>%
      select(-contains("prev_earinfection")) %>%
      select(-contains("MacArthur")) %>%
      select(-contains("method")) %>% # This is from MB1
      rename_all(recode, race_ethinicty = "race_ethnicity", Notes = "notes")
  
    
pd_raw <- full_join(pd_raw, temp_dataset)
    rm(temp_dataset)
  }
  
  # if the merged dataset doesn't exist, create it
  if (!exists("pd_raw")){
      pd_raw <- read_excel(file, col_names=TRUE) %>%
      mutate_all(as.character) %>%
      mutate_all(nc_na) %>%
      select(-(starts_with("Optional"))) %>%
      select(-(starts_with("parent")))%>%
      select(-contains("NAE")) %>%
      select(-starts_with("Bilingual")) %>%
      select(-contains("prev_earinfection")) %>%
      select(-contains("MacArthur")) %>%
      select(-contains("method")) %>% # This is from MB1
      rename_all(recode, race_ethinicty = "race_ethnicity", Notes = "notes")
  
  }
  
}





```


```{r}
pd_raw %>%
  vis_dat() + 
  coord_flip()
```

```{r}
n_participant_rows = nrow(pd_raw)
n_unique_labs_from_participants = pd_raw %>%
  summarize(labs = n_distinct(lab))
unique_participants_by_lab = pd_raw %>%
  group_by(lab) %>%
  summarize(participants = n_distinct(subid))
```

Right after trial import, there are `r n_participant_rows` rows in the pd_raw dataframe; this includes data from `r n_unique_labs_from_participants$labs` labs. The number of participants coming from each lab is as follows:

```{r}
print(unique_participants_by_lab, n=1e4)
```

# Gazepath Data Import

```{r}
gazepath_files <- dir("processed_data/data_gazepath/", pattern = ".xlsx", all.files = FALSE)
gazepath_files <- paste0("processed_data/data_gazepath/", gazepath_files) # Adds path to names


for (file in gazepath_files){
  
  # if the merged dataset does exist, append to it
  if (exists("gp_raw")){
    temp_dataset <- read_excel(file, col_names=TRUE) %>%
     mutate_all(as.character)
    
gp_raw <- full_join(gp_raw, temp_dataset)
    rm(temp_dataset)
  }
  
  # if the merged dataset doesn't exist, create it
  if (!exists("gp_raw")){
       gp_raw <- read_excel(file, col_names=TRUE) %>%
       mutate_all(as.character)
  
  }
  
}

length(unique(gp_raw$lab))

```

There are `r length(unique(gp_raw$lab))` lab ids that contributed eyetracking data. Note that here, koku-hamburg has two different ids.  Will be handled in a later section

# Handcoded Data Import

```{r}


handcoded_files <- dir("processed_data/handcoded_cleaned/", pattern = ".xlsx", all.files = FALSE)
handcoded_files <- paste0("processed_data/handcoded_cleaned/", handcoded_files) # Adds path to names


for (file in handcoded_files){
  
  # if the merged dataset does exist, append to it
  if (exists("hc_raw")){
    temp_dataset <- read_excel(file, col_names=TRUE) %>%
      mutate_all(as.character) %>%
      mutate_all(nc_na) %>%
      mutate(subid = as.character(subid)) %>%
      select(-contains("Note")) %>%
      {if("fixation_incongrent" %in% names(.)) rename(., fixation_incongruent=fixation_incongrent) else .} #https://stackoverflow.com/questions/34275576/avoiding-error-when-using-rename-in-dplyr-and-column-doesnt-exist
    
hc_raw <- full_join(hc_raw, temp_dataset)
    rm(temp_dataset)
  }
  
  # if the merged dataset doesn't exist, create it
  if (!exists("hc_raw")){
       hc_raw <- read_excel(file, col_names=TRUE) %>%
      mutate_all(as.character) %>%
      mutate_all(nc_na) %>%
      mutate(subid = as.character(subid)) %>%
      select(-contains("Note")) %>%
      {if("fixation_incongrent" %in% names(.)) rename(., fixation_incongruent=fixation_incongrent) else .}
  
  }
  
}

length(unique(hc_raw$lab))


```

There are `r length(unique(hc_raw$lab))` labs that contributed hand-coded data.  This includes on lab that also contributed eye-tracked data.

Join to get full dataset
```{r}

# Check if there are any repeated participants across hand coded and eyetracked data.  There are none
double_coded <- inner_join(gp_raw, hc_raw, by = c("lab", "subid"))

# Coding type

gp_raw$coding <- "gazepath"
hc_raw$coding <- "handcoding"

# Assume gazepath trials do not have an error

gp_raw$trial_error[is.na(gp_raw$trial_error)] <- "noerror"

# So join hand-coded and gazepath datasets  
td_raw <- full_join(gp_raw, hc_raw)

```

```{r}

td <- td_raw %>%
  # Change variables to correct types
  mutate_if(is.character, tolower) %>%
  mutate_at(vars(study_order, trial_num, latency, n_shift_congruent, n_shift_incongruent, fixation_congruent, fixation_incongruent), as.numeric)
#  mutate_at(vars(lab, subid, first_shift, trial_error, trial_error_type, coding), as.factor)
  

# Fix video names
    td$stimulus = gsub("\\..*","", td$stimulus)
    td$stimulus = gsub("X", "", td$stimulus)

pd <- pd_raw %>%
  mutate_if(is.character, tolower)
 # mutate_at(vars(lab, subid, trial_order, session_error), as.factor)
        

```

See if lab ids are consistent, and if not fix them

```{r}

# Trials dataframe
td$lab[td$lab == "koku-hamburg_12-15mos"] <- "koku-hamburg"
td$lab[td$lab == "koku-hamburg_6-9mos"] <- "koku-hamburg"
td$lab[td$lab == "upf-barcelona"] <- "upf_barcelona"
td$lab[td$lab == "nus"] <- "nusinfantlanguagecentre"
td$lab[td$lab == "brookes-babylab"] <- "babylab-brookes"
td$lab[td$lab == "lll"] <- "lll-liv"
td$lab[td$lab == "zurich"] <- "weltentdecker-zurich"
td$lab[td$lab == "princeton-babylab"] <- "babylab-princeton"
td$lab[td$lab == "westernsydney-babylab"] <- "babylab-westernsydney"
td$lab[td$lab == "babylab_westernsydney"] <- "babylab-westernsydney"
td$lab[td$lab == "wsu: kw"] <- "babylab-westernsydney"

# Participants dataframe
pd$lab[pd$lab == "wsu: kw"] <- "babylab-westernsydney"
pd$lab[pd$lab == "babylab_westernsydney"] <- "babylab-westernsydney"
pd$lab[pd$lab == "koku_hamburg"] <- "koku-hamburg"
pd$lab[pd$lab == "princeton-babylab"] <- "babylab-princeton"


setdiff(unique(pd$lab), unique(td$lab))
setdiff(unique(td$lab), unique(pd$lab))
         

         
```

Correcting lab-specific subid issues and row issues

```{r}

# remove weird row with zero latency but no other info
td <- td %>%
  filter(!is.na(lab) & !is.na(subid) & !is.na(trial_num))


#irl-concordia trials
td$subid[td$lab == "irl-concordia" & td$subid == "gf_s30_47622"] <- "gf12_s30_47622" # correcting typo
td$subid[td$lab == "irl-concordia" & td$subid == "gf16_s19_46048"] <- "gf12_s19_46048" # correcting typo
td$subid[td$lab == "irl-concordia"] <- str_extract(td$subid[td$lab == "irl-concordia"], "gf[[:digit:]]+_s[[:digit:]]+")

#irl-concordia participants
pd$subid[pd$lab == "irl-concordia" & pd$age_days <= 300] <- paste0("gf6_", pd$subid[pd$lab == "irl-concordia" & pd$age_days <= 300])
pd$subid[pd$lab == "irl-concordia" & pd$age_days >= 300] <- paste0("gf12_", pd$subid[pd$lab == "irl-concordia" & pd$age_days >= 300])

#babylab-brookes
td$subid[td$lab == "babylab-brookes"] <- sub("g0", "g", td$subid[td$lab == "babylab-brookes"]) # Not all data zero-padded

#babylab-westernsydney trials
td$subid[td$lab == "babylab-westernsydney"] <- sub("mb00", "", td$subid[td$lab == "babylab-westernsydney"])
td$subid[td$lab == "babylab-westernsydney"] <- sub("mb0", "", td$subid[td$lab == "babylab-westernsydney"])
td$subid[td$lab == "babylab-westernsydney"] <- sub("mb", "", td$subid[td$lab == "babylab-westernsydney"])

#babylab-westernsydney participants
pd$subid[pd$lab == "babylab-westernsydney"] <- sub("mb00", "", pd$subid[pd$lab == "babylab-westernsydney"])
pd$subid[pd$lab == "babylab-westernsydney"] <- sub("mb0", "", pd$subid[pd$lab == "babylab-westernsydney"])
pd$subid[pd$lab == "babylab-westernsydney"] <- sub("mb", "", pd$subid[pd$lab == "babylab-westernsydney"])

#babylab-brookes
pd$subid[pd$lab == "babylab-brookes"] <- sub("g0", "g", pd$subid[pd$lab == "babylab-brookes"]) # Not all data zero-padded


```



# Pre-merge Checking

```{r}

participants_premerge_td <- td %>%
  group_by(lab, subid) %>%
  summarize(trialcount = n_distinct(trial_num), 
            rowcount = length(trial_num), trial_error = last(trial_error_type),present_td=TRUE)

participants_premerge_pd <- pd %>%
  group_by(lab, subid) %>%
  summarize(lines_in_pd = n(), age_days = first(age_days), notes = first(notes),days_preterm = first(days_preterm), session_error = first(session_error), session_error_type = first(session_error_type), notes = first(notes), present_pd=TRUE)

all_participants_premerge <- merge(participants_premerge_td, participants_premerge_pd, all.x=TRUE, all.y=TRUE) %>%
  replace_na(list(present_pd = FALSE, present_td=FALSE))

unmatched_participants_premerge <- filter(all_participants_premerge, !(present_td & present_pd)) %>%
  filter(session_error == "noerror")

# Check if we have resolved any of these issues

conf <- read_csv('metadata/true_unmatched_participants.csv') %>%
  filter(confirmed == 'X')%>%
  select(subid, lab, confirmed)

unmatched_participants_premerge <- merge(unmatched_participants_premerge, conf, all.x=TRUE) %>%
  replace_na(list(confirmed = FALSE))%>%
  filter(confirmed != 'X')

write_csv(unmatched_participants_premerge, 'metadata/unconfirmed_unmatched_participants.csv')

unmatched_participants_premerge
 
 # participants_session_errors_all <- pd %>% filter(session_error!= "noerror") %>%
 #   select(lab, subid, session_error, session_error_type, notes) %>%
 #   filter(session_error != "na") %>%
 #   unique()
 # 
 # write_csv(participants_session_errors_all, 'metadata/participants_session_errors_all.csv')


# participants_trial_errors_all <- td %>% filter(trial_error == "error")
# 
# write_csv(participants_trial_errors_all, 'metadata/participants_trial_errors_all.csv')

```

We still have a `r length(unmatched_participants_premerge)` that need to be checked on.



```{r}

```



# Merge

Use `inner_join` to get matching participants. Our process is then to use testing to identify where we are missing participants in this join. 

```{r}
d <- inner_join(td, pd) 
```


# Test Merge

```{r}
participants_postmerge <- d %>%
  group_by(lab, subid) %>%
  summarize(trialcount = n_distinct(trial_num), 
            rowcount = length(trial_num))
```

Compare the number of rows and trials to pre-merge values! In the resulting dataframe, we have `r nrow(setdiff(select(participants_premerge_td, lab, subid, trialcount, rowcount), participants_postmerge))` participants missing from TD and `r nrow(setdiff(select(participants_premerge_pd, lab, subid), select(participants_postmerge, lab, subid)))` from PD.  Where did they go, and why did they drop?

Here's a full list of participants (and labs) who get lost during merge. Use `anti_join` to detect lost participants. (For sanity, this should be the same as the pre-merged list of 'problem' subids)

```{r}
lost_participants_td <- anti_join(participants_premerge_td, participants_postmerge)
lost_participants_pd <- anti_join(participants_premerge_pd, participants_postmerge)


validate_that(nrow(lost_participants_td) + nrow(lost_participants_pd) == nrow(unmatched_participants_premerge))

lost_participants_td
lost_participants_pd

```


# Output

Output intermediate files. 

```{r}
write_csv(d, "processed_data/01_merged_output.csv")
```
