==> 03_exclusion.Rmd <==
---
title: "Gaze Following Exclusions"
author: Gaze following Analysis Team"
date: '`r format(Sys.time(), "%a %b %d %X %Y")`'
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: yes
---


# Intro

This script implements and documents exclusion criteria. 

```{r setup, echo=FALSE, message=FALSE}
source(here::here("helper/common.R"))
source(here("helper/preprocessing_helper.R"))
```

# Exclusions

Note that all exclusions are written in paper format and sourced here so as to allow matching exactly to the data exclusion script. 

```{r child="paper/exclusions.Rmd"}

source_rmd("paper/exclusions.Rmd")

```


## Blinding 

Current data output is **UNBLINDED.** 

```{r}
#d <- d %>%
#   group_by(lab, subid, stimulus_num) %>%
#   mutate(trial_type = base::sample(trial_type))
```



and write: 

```{r}
write_csv(d, "processed_data/03_data_trial_main.csv")
```
